Refactored consumption int into long
Questions should now be able to display the whole consumption value and prevent overflow from happening

+++ client/src/main/java/client/scenes/MultiplayerScreenCtrl.java
@@ -185,7 +185,7 @@ public class MultiplayerScreenCtrl {
             if(a == null) continue;
             optionToID.put(options.get(i), a.id);
             
-            titles.get(i).setText(Integer.toString(a.consumptionInWh));
+            titles.get(i).setText(Long.toString(a.consumptionInWh));
             descriptions.get(i).setText(a.title);
 
             images.get(i).setImage(new Image("http://localhost:8080/activities/" + a.imagePath));
