+++ client/src/main/java/client/scenes/SingleplayerScreenCtrl.java
@@ -131,7 +131,7 @@ public class SingleplayerScreenCtrl {
 
             optionToID.put(options.get(i), a.id);
 
-            titles.get(i).setText(Integer.toString(a.consumptionInWh));
+            titles.get(i).setText(Long.toString(a.consumptionInWh));
             descriptions.get(i).setText(a.title);
 
             images.get(i).setImage(new Image("http://localhost:8080/activities/" + a.imagePath));
